fahr to celsius 浮点数版本

+++ chapter1/bookcase/fahr2celsius.c
@@ -1,9 +1,10 @@
 //"华氏度和摄氏度温度对照表"
+//"浮点数版本"
 #include <stdio.h>
 
 int main()
 {
-	int fahr, celsius;
+	float fahr, celsius;
 	int lower, upper, step;
 
 	lower = 0;
@@ -12,8 +13,8 @@ int main()
 
 	fahr = lower;
 	while (fahr <= upper) {
-		celsius = 5 * (fahr - 32) / 9;
-		printf("%3d\t%3d\n", fahr, celsius);
+		celsius = 5.0 * (fahr - 32.0) / 9.0;
+		printf("%3.0f\t%5.2f\n", fahr, celsius);
 		fahr += step;
 	}
 	return 0;
